requirements.txt
Flask>=2.0
Flask-SQLAlchemy>=3.0
Flask-Migrate>=4.0
Flask-JWT-Extended>=4.0
python-dotenv>=1.0
gunicorn>=20.0
psycopg2-binary>=2.9  # optional for Postgres

# Websocket / realtime
Flask-SocketIO>=5.0
python-socketio>=5.0
python-engineio>=4.0
eventlet>=0.33  # or gevent if preferred

# embeddings / vector store
sentence-transformers>=2.2.2
faiss-cpu==1.12.0

# LangChain + HF embeddings for simpler FAISS integration
langchain==0.1.0
transformers==4.36.0
huggingface-hub>=0.13.0

# AWS / HTTP
boto3>=1.26
requests>=2.0
python-docx>=0.8.11
# optional: textract for legacy .doc support (install if needed)
# textract>=1.6.5
